Refactored new interface and struct-utils file
